Add files via upload
The program is to find the middle of link list using slow and fast pointer approach. it will be helpful if you don't want to traverse the whole list for counting length of link list and later on finding the middle value. This will operate in O(n) time complexity and other than the size of link list it will   require  just two pointer variables only so space complexity will be O(n) .